🚚 Move hidden and featured endpoints to `/users` folder
