+++ src/server/api/routes/users/v2/nfts/hidden.js
@@ -1,16 +1,16 @@
 const { Router } = require('express');
-const { authenticateLogin } = require('../../middleware/auth');
-const { setPrivateCacheHeader } = require('../../middleware/cache');
-const { handleRestfulError } = require('../../middleware/error');
-const { isValidAddress } = require('../../util/cosmos');
+const { authenticateLogin } = require('../../../../middleware/auth');
+const { setPrivateCacheHeader } = require('../../../../middleware/cache');
+const { handleRestfulError } = require('../../../../middleware/error');
+const { isValidAddress } = require('../../../../util/cosmos');
 const {
   FieldValue,
   walletUserCollection,
-} = require('../../../modules/firebase');
+} = require('../../../../../modules/firebase');
 
 const router = Router();
 
-router.get('/nft/featured', async (req, res, next) => {
+router.get('/nft/hidden', async (req, res, next) => {
   try {
     const userId = req.query.owner;
     if (!isValidAddress(userId)) {
@@ -18,14 +18,14 @@ router.get('/nft/featured', async (req, res, next) => {
       return;
     }
     const userDoc = await walletUserCollection(userId).get();
-    const { featuredNftClassIds = [] } = userDoc.data();
-    res.json({ featured: featuredNftClassIds });
+    const { hiddenNftClassIds = [] } = userDoc.data();
+    res.json({ hidden: hiddenNftClassIds });
   } catch (err) {
     handleRestfulError(req, res, next, err);
   }
 });
 
-router.post('/nft/featured', authenticateLogin, async (req, res, next) => {
+router.post('/nft/hidden', authenticateLogin, async (req, res, next) => {
   try {
     setPrivateCacheHeader(res);
     const { user } = req.session;
@@ -35,7 +35,7 @@ router.post('/nft/featured', authenticateLogin, async (req, res, next) => {
       return;
     }
     await walletUserCollection.doc(user).update({
-      featuredNftClassIds: FieldValue.arrayUnion(...nftClassIds),
+      hiddenNftClassIds: FieldValue.arrayUnion(...nftClassIds),
     });
     res.sendStatus(200);
   } catch (err) {
@@ -43,7 +43,7 @@ router.post('/nft/featured', authenticateLogin, async (req, res, next) => {
   }
 });
 
-router.delete('/nft/featured', authenticateLogin, async (req, res, next) => {
+router.delete('/nft/hidden', authenticateLogin, async (req, res, next) => {
   try {
     setPrivateCacheHeader(res);
     const { user } = req.session;
@@ -53,7 +53,7 @@ router.delete('/nft/featured', authenticateLogin, async (req, res, next) => {
       return;
     }
     await walletUserCollection.doc(user).update({
-      featuredNftClassIds: FieldValue.arrayRemove(...nftClassIds),
+      hiddenNftClassIds: FieldValue.arrayRemove(...nftClassIds),
     });
     res.sendStatus(200);
   } catch (err) {
