+++ src/server/api/routes/users/v2/nfts/index.js
@@ -1,0 +1,11 @@
+const { Router } = require('express');
+
+const featured = require('./featured');
+const hidden = require('./hidden');
+
+const router = Router();
+
+router.use(featured);
+router.use(hidden);
+
+module.exports = router;
